requirements.txt
apache-airflow[gcp_api,postgres]==1.10.3
